refactor: split olf.e2e into runner/health/dagster/trino/layers/assertions/artifacts/preflight
Decompose the 1094-line flat olf/e2e.py into a package split by
capability, per issue #101: _shell (config/error type and
process/kubectl primitives -- the acyclic leaf every other submodule
depends on), _runner (environment setup, suite dispatch, full-suite
assertion inventory -- the composition root), _health, _dagster,
_trino, _layers (optional governance/analytics layer gating),
_assertions (OpenMetadata/Superset), _artifacts, and _preflight (AWS
provider checks). load_provider_contracts_or_raise and aws_stack_region
moved into _shell rather than _preflight since they're consumed by
_layers and _artifacts too, not just AWS preflight -- keeping them on
the leaf module avoids circular imports.

Reorganized the 1588-line test_e2e.py into one file per submodule
(test_e2e_runner.py etc.), following the existing test_<package>_
<module>.py convention. Introduced tests/conftest.py for the E2EConfig
builder and inventory fixtures shared across all eight files. Where a
test's target function calls through a name re-imported into another
module (e.g. _runner calling _trino/_dagster/_assertions/_artifacts/
_preflight functions, or _run_retry in _shell calling _shell's own
_run), monkeypatches target the module that actually performs the
call, not the module where the symbol was originally defined.

Also updates scripts/test/check-structure.sh's registered paths for
e2e.py and openmetadata.py to their new __init__.py locations.

No CLI behavior change. make local-up/local-e2e were not run: this
environment's Docker allocation (7.7GB/4 CPUs) is insufficient for the
full stack and local-up OOM-killed (exit 137) in olf/superset.py's pod
exec during report deploy -- unrelated code, not touched by this
change. All static gates pass: check-structure, check-components,
check-contracts, check-infra, check-project-code, check-dbt,
check-lockfiles, ruff, and pytest (302 tests). CLI command
names/flags/help are unchanged (verified via `olf --help`,
`olf e2e/openmetadata/release --help`).
