+++ scripts/test/check-structure.sh
@@ -153,10 +153,10 @@ required_paths=(
   "packages/domain-model/openlakeforge_domain/inventory.py"
   "tools/olf/olf/floe.py"
   "tools/olf/olf/k8s.py"
-  "tools/olf/olf/e2e.py"
+  "tools/olf/olf/e2e/__init__.py"
   "tools/olf/olf/s3.py"
   "tools/olf/olf/superset.py"
-  "tools/olf/olf/openmetadata.py"
+  "tools/olf/olf/openmetadata/__init__.py"
   "scripts/artifacts/floe-manifest.sh"
   "scripts/artifacts/dbt-parse.sh"
   "scripts/artifacts/olf.sh"
