+++ tools/olf/tests/test_e2e_layers.py
@@ -1,0 +1,16 @@
+from pathlib import Path
+
+import pytest
+from conftest import e2e_cfg
+
+from olf.e2e import _layers
+
+
+def test_configured_layers_reads_contract_flags(monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> None:
+    monkeypatch.setattr(
+        _layers,
+        "load_provider_contracts_or_raise",
+        lambda _cfg: {"governance": {"enabled": False}, "reporting": {"enabled": False}},
+    )
+
+    assert _layers.configured_layers(e2e_cfg(tmp_path)) == {"governance": False, "analytics": False}
